#1467 - Modularize agreement
- Deprecating outdates methods and classes
